+++ src/apis/api.module.ts
@@ -1,4 +1,6 @@
 import { Module } from '@nestjs/common';
+import { FreePostCommentsModule } from '@src/apis/free-post-comments/free-post-comments.module';
+import { FreePostReplyCommentsModule } from '@src/apis/free-post-reply-comments/free-post-reply-comments.module';
 import { RootModule } from '@src/apis/root/root.module';
 import { AuthModule } from './auth/auth.module';
 import { FreePostsModule } from './free-posts/free-posts.module';
@@ -14,6 +16,8 @@ import { UsersModule } from './users/users.module';
     FreePostsModule,
     MajorModule,
     NoticePostsModule,
+    FreePostCommentsModule,
+    FreePostReplyCommentsModule,
   ],
 })
 export class ApiModule {}
