+++ src/apis/common-posts/common-posts.module.ts
@@ -2,6 +2,7 @@ import { DynamicModule, Module, Type } from '@nestjs/common';
 import { COMMON_POST_REPOSITORY_TOKEN } from '@src/apis/common-posts/constants/common-posts.token';
 import { CommonPostsService } from '@src/apis/common-posts/services/common-posts.service';
 import { RequiredCommonPostColumn } from '@src/apis/common-posts/types/common-post.type';
+import { DataSource } from 'typeorm';
 
 @Module({
   providers: [CommonPostsService],
@@ -14,7 +15,10 @@ export class CommonPostsModule {
       providers: [
         {
           provide: COMMON_POST_REPOSITORY_TOKEN,
-          useClass: postEntity,
+          useFactory: (dataSource: DataSource) => {
+            return dataSource.getRepository(postEntity);
+          },
+          inject: [DataSource],
         },
       ],
     };
