+++ src/apis/free-post-comments/dto/find-free-post-comment-list-query.dto.ts
@@ -1,6 +1,6 @@
-import { FreePostCommentStatus } from '@src/apis/free-posts/constants/free-post-comment.enum';
+import { FreePostCommentStatus } from '@src/apis/free-post-comments/constants/free-post-comment.enum';
+import { FreePostCommentDto } from '@src/apis/free-post-comments/dto/free-post-comment.dto';
 import { FREE_POST_ORDER_FIELD } from '@src/apis/free-posts/constants/free-post.constant';
-import { FreePostCommentDto } from '@src/apis/free-posts/dto/free-post-comment.dto';
 import { SortOrder } from '@src/constants/enum';
 import { PageDto } from '@src/dto/page.dto';
 import { ApiPropertyOrder } from '@src/dto/swagger/api-property-order.decorator';
