+++ src/apis/free-post-comments/dto/put-update-free-post-comment.dto.ts
@@ -1,5 +1,5 @@
 import { ApiProperty, PickType } from '@nestjs/swagger';
-import { CreateFreePostCommentDto } from '@src/apis/free-posts/dto/create-free-post-comment.dto';
+import { CreateFreePostCommentDto } from '@src/apis/free-post-comments/dto/create-free-post-comment.dto';
 import { IsBoolean } from 'class-validator';
 
 export class PutUpdateFreePostCommentDto extends PickType(
