+++ src/apis/free-post-reply-comments/controllers/free-post-reply-comments.swagger.ts
@@ -1,13 +1,9 @@
 import { HttpStatus, applyDecorators } from '@nestjs/common';
-import {
-  ApiBearerAuth,
-  ApiNoContentResponse,
-  ApiOperation,
-} from '@nestjs/swagger';
+import { ApiBearerAuth, ApiOperation } from '@nestjs/swagger';
 import { OperationObject } from '@nestjs/swagger/dist/interfaces/open-api-spec.interface';
-import { FreePostsController } from '@src/apis/free-posts/controllers/free-posts.controller';
-import { FreePostDto } from '@src/apis/free-posts/dto/free-post.dto';
-import { FreePostsItemDto } from '@src/apis/free-posts/dto/free-posts-item.dto';
+import { FreePostReplyCommentsController } from '@src/apis/free-post-reply-comments/controllers/free-post-reply-comments.controller';
+import { FreePostReplyCommentDto } from '@src/apis/free-post-reply-comments/dto/free-post-reply-comment.dto';
+import { FreePostReplyCommentsItemDto } from '@src/apis/free-post-reply-comments/dto/free-post-reply-comments-item.dto';
 import { COMMON_ERROR_CODE } from '@src/constants/error/common/common-error-code.constant';
 import { HttpException } from '@src/http-exceptions/exceptions/http.exception';
 import { DeleteResponseDto } from '@src/interceptors/success-interceptor/dto/delete-response.dto';
@@ -19,21 +15,23 @@ import { ValidationError } from '@src/types/validation-errors.type';
 /**
  * @todo
  */
-export const ApiFreePost: ApiOperator<keyof FreePostsController> = {
+export const ApiFreePostReplyComment: ApiOperator<
+  keyof FreePostReplyCommentsController
+> = {
   Create: (
     apiOperationOptions: Required<Pick<Partial<OperationObject>, 'summary'>> &
       Partial<OperationObject>,
   ): PropertyDecorator => {
     return applyDecorators(
       ApiOperation({
-        operationId: 'FreePostCreate',
+        operationId: 'FreePostReplyCommentCreate',
         ...apiOperationOptions,
       }),
       ApiBearerAuth(),
       DetailResponseDto.swaggerBuilder(
         HttpStatus.CREATED,
-        'freePost',
-        FreePostDto,
+        'freePostReplyComment',
+        FreePostReplyCommentDto,
       ),
       HttpException.swaggerBuilder(
         HttpStatus.BAD_REQUEST,
@@ -46,6 +44,9 @@ export const ApiFreePost: ApiOperator<keyof FreePostsController> = {
       ),
       HttpException.swaggerBuilder(HttpStatus.UNAUTHORIZED, [
         COMMON_ERROR_CODE.INVALID_TOKEN,
+      ]),
+      HttpException.swaggerBuilder(HttpStatus.NOT_FOUND, [
+        COMMON_ERROR_CODE.RESOURCE_NOT_FOUND,
       ]),
       HttpException.swaggerBuilder(HttpStatus.INTERNAL_SERVER_ERROR, [
         COMMON_ERROR_CODE.SERVER_ERROR,
@@ -59,13 +60,13 @@ export const ApiFreePost: ApiOperator<keyof FreePostsController> = {
   ): PropertyDecorator => {
     return applyDecorators(
       ApiOperation({
-        operationId: 'FreePostFindAllAndCount',
+        operationId: 'FreePostReplyCommentFindAllAndCount',
         ...apiOperationOptions,
       }),
       PaginationResponseDto.swaggerBuilder(
         HttpStatus.OK,
-        'freePosts',
-        FreePostsItemDto,
+        'freePostReplyComments',
+        FreePostReplyCommentsItemDto,
       ),
       HttpException.swaggerBuilder(
         HttpStatus.BAD_REQUEST,
@@ -79,28 +80,6 @@ export const ApiFreePost: ApiOperator<keyof FreePostsController> = {
       HttpException.swaggerBuilder(HttpStatus.INTERNAL_SERVER_ERROR, [
         COMMON_ERROR_CODE.SERVER_ERROR,
       ]),
-    );
-  },
-
-  FindOneOrNotFound: (
-    apiOperationOptions: Required<Pick<Partial<OperationObject>, 'summary'>> &
-      Partial<OperationObject>,
-  ): PropertyDecorator => {
-    return applyDecorators(
-      ApiOperation({
-        operationId: 'FreePostFindOneOrNotFound',
-        ...apiOperationOptions,
-      }),
-      DetailResponseDto.swaggerBuilder(HttpStatus.OK, 'freePost', FreePostDto),
-      HttpException.swaggerBuilder(
-        HttpStatus.BAD_REQUEST,
-        [COMMON_ERROR_CODE.INVALID_REQUEST_PARAMETER],
-        {
-          description:
-            '해당 필드는 request parameter 가 잘못된 경우에만 리턴됩니다.',
-          type: ValidationError,
-        },
-      ),
       HttpException.swaggerBuilder(HttpStatus.NOT_FOUND, [
         COMMON_ERROR_CODE.RESOURCE_NOT_FOUND,
       ]),
@@ -113,52 +92,18 @@ export const ApiFreePost: ApiOperator<keyof FreePostsController> = {
   ): PropertyDecorator => {
     return applyDecorators(
       ApiOperation({
-        operationId: 'FreePostPutUpdate',
+        operationId: 'FreePostReplyCommentPutUpdate',
         ...apiOperationOptions,
       }),
       ApiBearerAuth(),
-      DetailResponseDto.swaggerBuilder(HttpStatus.OK, 'freePost', FreePostDto),
+      DetailResponseDto.swaggerBuilder(
+        HttpStatus.OK,
+        'freePostReplyComment',
+        FreePostReplyCommentDto,
+      ),
       HttpException.swaggerBuilder(
         HttpStatus.BAD_REQUEST,
         [COMMON_ERROR_CODE.INVALID_REQUEST_PARAMETER],
-        {
-          description:
-            '해당 필드는 request parameter 가 잘못된 경우에만 리턴됩니다.',
-          type: ValidationError,
-        },
-      ),
-      HttpException.swaggerBuilder(HttpStatus.UNAUTHORIZED, [
-        COMMON_ERROR_CODE.INVALID_TOKEN,
-      ]),
-      HttpException.swaggerBuilder(HttpStatus.FORBIDDEN, [
-        COMMON_ERROR_CODE.PERMISSION_DENIED,
-      ]),
-      HttpException.swaggerBuilder(HttpStatus.NOT_FOUND, [
-        COMMON_ERROR_CODE.RESOURCE_NOT_FOUND,
-      ]),
-      HttpException.swaggerBuilder(HttpStatus.INTERNAL_SERVER_ERROR, [
-        COMMON_ERROR_CODE.SERVER_ERROR,
-      ]),
-    );
-  },
-
-  PatchUpdate: (
-    apiOperationOptions: Required<Pick<Partial<OperationObject>, 'summary'>> &
-      Partial<OperationObject>,
-  ): PropertyDecorator => {
-    return applyDecorators(
-      ApiOperation({
-        operationId: 'FreePostPatchUpdate',
-        ...apiOperationOptions,
-      }),
-      ApiBearerAuth(),
-      DetailResponseDto.swaggerBuilder(HttpStatus.OK, 'freePost', FreePostDto),
-      HttpException.swaggerBuilder(
-        HttpStatus.BAD_REQUEST,
-        [
-          COMMON_ERROR_CODE.INVALID_REQUEST_PARAMETER,
-          COMMON_ERROR_CODE.MISSING_UPDATE_FIELD,
-        ],
         {
           description:
             '해당 필드는 request parameter 가 잘못된 경우에만 리턴됩니다.',
@@ -186,7 +131,7 @@ export const ApiFreePost: ApiOperator<keyof FreePostsController> = {
   ): PropertyDecorator => {
     return applyDecorators(
       ApiOperation({
-        operationId: 'FreePostRemove',
+        operationId: 'FreePostReplyCommentRemove',
         ...apiOperationOptions,
       }),
       ApiBearerAuth(),
@@ -211,31 +156,6 @@ export const ApiFreePost: ApiOperator<keyof FreePostsController> = {
       ]),
       HttpException.swaggerBuilder(HttpStatus.INTERNAL_SERVER_ERROR, [
         COMMON_ERROR_CODE.SERVER_ERROR,
-      ]),
-    );
-  },
-
-  IncrementHit: function (
-    apiOperationOptions: Required<Pick<Partial<OperationObject>, 'summary'>> &
-      Partial<OperationObject>,
-  ): PropertyDecorator {
-    return applyDecorators(
-      ApiOperation({
-        operationId: 'FreePostIncrementHit',
-        ...apiOperationOptions,
-      }),
-      ApiNoContentResponse(),
-      HttpException.swaggerBuilder(
-        HttpStatus.BAD_REQUEST,
-        [COMMON_ERROR_CODE.INVALID_REQUEST_PARAMETER],
-        {
-          description:
-            '해당 필드는 request parameter 가 잘못된 경우에만 리턴됩니다.',
-          type: ValidationError,
-        },
-      ),
-      HttpException.swaggerBuilder(HttpStatus.NOT_FOUND, [
-        COMMON_ERROR_CODE.RESOURCE_NOT_FOUND,
       ]),
     );
   },
