+++ src/apis/free-post-reply-comments/dto/put-update-free-post-reply-comment.dto.ts
@@ -1,5 +1,5 @@
 import { ApiProperty, PickType } from '@nestjs/swagger';
-import { CreateFreePostReplyCommentDto } from '@src/apis/free-posts/dto/create-free-post-reply-comment.dto';
+import { CreateFreePostReplyCommentDto } from '@src/apis/free-post-reply-comments/dto/create-free-post-reply-comment.dto';
 import { IsBoolean } from 'class-validator';
 
 export class PutUpdateFreePostReplyCommentDto extends PickType(
