+++ src/apis/free-post-reply-comments/free-post-reply-comments.module.ts
@@ -1,0 +1,19 @@
+import { Module } from '@nestjs/common';
+import { FreePostCommentsModule } from '@src/apis/free-post-comments/free-post-comments.module';
+import { FreePostReplyCommentHistoryModule } from '@src/apis/free-post-reply-comments/free-post-reply-comment-history/free-post-reply-comment-history.module';
+import { FreePostReplyCommentRepository } from '@src/apis/free-posts/repositories/free-post-reply-comment.repository';
+import { TypeOrmExModule } from '@src/core/type-orm/type-orm-ex.module';
+import { QueryHelper } from '@src/helpers/query.helper';
+import { FreePostReplyCommentsController } from './controllers/free-post-reply-comments.controller';
+import { FreePostReplyCommentsService } from './services/free-post-reply-comments.service';
+
+@Module({
+  imports: [
+    FreePostCommentsModule,
+    FreePostReplyCommentHistoryModule,
+    TypeOrmExModule.forCustomRepository([FreePostReplyCommentRepository]),
+  ],
+  controllers: [FreePostReplyCommentsController],
+  providers: [FreePostReplyCommentsService, QueryHelper],
+})
+export class FreePostReplyCommentsModule {}
