--- src/apis/free-posts/dto/free-post-reply-comments-item.dto.ts
@@ -1,3 +1,0 @@
-import { FreePostReplyCommentDto } from '@src/apis/free-posts/dto/free-post-reply-comment.dto';
-
-export class FreePostReplyCommentsItemDto extends FreePostReplyCommentDto {}
