+++ src/apis/free-posts/free-post-history/free-post-history.module.ts
@@ -1,7 +1,7 @@
 import { Module } from '@nestjs/common';
-import { FreePostCommentHistoryRepository } from '@src/apis/free-posts/free-post-history/repositories/free-post-comment-history.repository';
+import { FreePostCommentHistoryRepository } from '@src/apis/free-post-comments/free-post-comment-history/repositories/free-post-comment-history.repository';
+import { FreePostReplyCommentHistoryRepository } from '@src/apis/free-post-reply-comments/free-post-reply-comment-history/repositories/free-post-reply-comment-history.repository';
 import { FreePostHistoryRepository } from '@src/apis/free-posts/free-post-history/repositories/free-post-history.repository';
-import { FreePostReplyCommentHistoryRepository } from '@src/apis/free-posts/free-post-history/repositories/free-post-reply-comment-history.repository';
 import { TypeOrmExModule } from '@src/core/type-orm/type-orm-ex.module';
 import { FreePostHistoryService } from './services/free-post-history.service';
