+++ src/apis/free-posts/free-post-history/services/free-post-history.service.spec.ts
@@ -1,16 +1,10 @@
 import { Test, TestingModule } from '@nestjs/testing';
-import { CreateFreePostCommentHistoryDto } from '@src/apis/free-posts/free-post-history/dto/create-free-post-comment-history.dto';
 import { CreateFreePostHistoryDto } from '@src/apis/free-posts/free-post-history/dto/create-free-post-history.dto';
-import { CreateFreePostReplyCommentHistoryDto } from '@src/apis/free-posts/free-post-history/dto/create-free-post-reply-comment-history.dto';
-import { FreePostCommentHistoryRepository } from '@src/apis/free-posts/free-post-history/repositories/free-post-comment-history.repository';
 import { FreePostHistoryRepository } from '@src/apis/free-posts/free-post-history/repositories/free-post-history.repository';
-import { FreePostReplyCommentHistoryRepository } from '@src/apis/free-posts/free-post-history/repositories/free-post-reply-comment-history.repository';
 import { HistoryAction } from '@src/constants/enum';
 import {
   mockEntityManager,
-  mockFreePostCommentHistoryRepository,
   mockFreePostHistoryRepository,
-  mockFreePostReplyCommentHistoryRepository,
 } from '@test/mock/mock.repository';
 import { FreePostHistoryService } from './free-post-history.service';
 
@@ -24,14 +18,6 @@ describe(FreePostHistoryService.name, () => {
         {
           provide: FreePostHistoryRepository,
           useValue: mockFreePostHistoryRepository,
-        },
-        {
-          provide: FreePostCommentHistoryRepository,
-          useValue: mockFreePostCommentHistoryRepository,
-        },
-        {
-          provide: FreePostReplyCommentHistoryRepository,
-          useValue: mockFreePostReplyCommentHistoryRepository,
         },
       ],
     }).compile();
@@ -79,73 +65,27 @@ describe(FreePostHistoryService.name, () => {
     });
   });
 
-  describe(FreePostHistoryService.prototype.createComment.name, () => {
+  describe(FreePostHistoryService.prototype.findOneOrFail.name, () => {
     let entityManager: any;
-    let userId: number;
-    let freePostId: number;
-    let action: HistoryAction;
-    let createFreePostCommentHistoryDto: CreateFreePostCommentHistoryDto;
+    let options: any;
 
     beforeEach(() => {
       entityManager = mockEntityManager;
-      userId = NaN;
-      freePostId = NaN;
-      action = null;
-      createFreePostCommentHistoryDto = new CreateFreePostCommentHistoryDto(
-        {} as any,
+      options = {};
+    });
+
+    it('find one or fail', async () => {
+      const newHistory = {};
+
+      mockFreePostHistoryRepository.findOneOrFail.mockResolvedValue(newHistory);
+
+      await expect(
+        service.findOneOrFail(entityManager, options),
+      ).resolves.toEqual(newHistory);
+
+      expect(mockFreePostHistoryRepository.findOneOrFail).toHaveBeenCalledWith(
+        options,
       );
-    });
-
-    it('create comment history', async () => {
-      mockFreePostHistoryRepository.findOneOrFail.mockResolvedValue({ id: 1 });
-      mockFreePostCommentHistoryRepository.save.mockResolvedValue({});
-
-      await expect(
-        service.createComment(
-          entityManager,
-          userId,
-          freePostId,
-          action,
-          createFreePostCommentHistoryDto,
-        ),
-      ).resolves.toEqual({});
-    });
-  });
-
-  describe(FreePostHistoryService.prototype.createReplyComment.name, () => {
-    let entityManager: any;
-    let userId: number;
-    let freePostId: number;
-    let action: HistoryAction;
-    let createFreePostReplyCommentHistoryDto: CreateFreePostReplyCommentHistoryDto;
-
-    beforeEach(() => {
-      entityManager = mockEntityManager;
-      userId = NaN;
-      freePostId = NaN;
-      action = null;
-      createFreePostReplyCommentHistoryDto =
-        new CreateFreePostReplyCommentHistoryDto({} as any);
-    });
-
-    it('create reply comment history', async () => {
-      mockFreePostHistoryRepository.findOneOrFail.mockResolvedValue({
-        id: 1,
-      });
-      mockFreePostCommentHistoryRepository.findOneOrFail.mockResolvedValue({
-        id: 1,
-      });
-      mockFreePostReplyCommentHistoryRepository.save.mockResolvedValue({});
-
-      await expect(
-        service.createReplyComment(
-          entityManager,
-          userId,
-          freePostId,
-          action,
-          createFreePostReplyCommentHistoryDto,
-        ),
-      ).resolves.toEqual({});
     });
   });
 });
