+++ src/apis/free-posts/free-post-history/services/free-post-history.service.ts
@@ -1,19 +1,14 @@
 import { Injectable } from '@nestjs/common';
-import { CreateFreePostCommentHistoryDto } from '@src/apis/free-posts/free-post-history/dto/create-free-post-comment-history.dto';
 import { CreateFreePostHistoryDto } from '@src/apis/free-posts/free-post-history/dto/create-free-post-history.dto';
-import { CreateFreePostReplyCommentHistoryDto } from '@src/apis/free-posts/free-post-history/dto/create-free-post-reply-comment-history.dto';
-import { FreePostCommentHistoryRepository } from '@src/apis/free-posts/free-post-history/repositories/free-post-comment-history.repository';
 import { FreePostHistoryRepository } from '@src/apis/free-posts/free-post-history/repositories/free-post-history.repository';
-import { FreePostReplyCommentHistoryRepository } from '@src/apis/free-posts/free-post-history/repositories/free-post-reply-comment-history.repository';
-import { HistoryAction, SortOrder } from '@src/constants/enum';
-import { EntityManager } from 'typeorm';
+import { HistoryAction } from '@src/constants/enum';
+import { FreePostHistory } from '@src/entities/FreePostHistory';
+import { EntityManager, FindOneOptions } from 'typeorm';
 
 @Injectable()
 export class FreePostHistoryService {
   constructor(
     private readonly freePostHistoryRepository: FreePostHistoryRepository,
-    private readonly freePostCommentHistoryRepository: FreePostCommentHistoryRepository,
-    private readonly freePostReplyCommentHistoryRepository: FreePostReplyCommentHistoryRepository,
   ) {}
 
   create(
@@ -31,78 +26,12 @@ export class FreePostHistoryService {
     });
   }
 
-  async createComment(
+  findOneOrFail(
     entityManager: EntityManager,
-    userId: number,
-    freePostId: number,
-    action: HistoryAction,
-    createFreePostCommentHistoryDto: CreateFreePostCommentHistoryDto,
+    options: FindOneOptions<FreePostHistory>,
   ) {
-    const recentFreePostHistory = await entityManager
-      .withRepository(this.freePostHistoryRepository)
-      .findOneOrFail({
-        select: {
-          id: true,
-        },
-        where: {
-          freePostId,
-        },
-        order: {
-          id: SortOrder.Desc,
-        },
-      });
-
     return entityManager
-      .withRepository(this.freePostCommentHistoryRepository)
-      .save({
-        userId,
-        action,
-        freePostHistoryId: recentFreePostHistory.id,
-        ...new CreateFreePostCommentHistoryDto(createFreePostCommentHistoryDto),
-      });
-  }
-
-  async createReplyComment(
-    entityManager: EntityManager,
-    userId: number,
-    freePostId: number,
-    action: HistoryAction,
-    createFreePostReplyCommentHistoryDto: CreateFreePostReplyCommentHistoryDto,
-  ) {
-    const recentFreePostHistory = await entityManager
       .withRepository(this.freePostHistoryRepository)
-      .findOneOrFail({
-        select: {
-          id: true,
-        },
-        where: {
-          freePostId,
-        },
-        order: {
-          id: SortOrder.Desc,
-        },
-      });
-
-    const recentFreePostCommentHistory = await entityManager
-      .withRepository(this.freePostCommentHistoryRepository)
-      .findOneOrFail({
-        select: {
-          id: true,
-        },
-        where: {
-          freePostHistoryId: recentFreePostHistory.id,
-        },
-      });
-
-    return entityManager
-      .withRepository(this.freePostReplyCommentHistoryRepository)
-      .save({
-        userId,
-        action,
-        freePostCommentHistoryId: recentFreePostCommentHistory.id,
-        ...new CreateFreePostReplyCommentHistoryDto(
-          createFreePostReplyCommentHistoryDto,
-        ),
-      });
+      .findOneOrFail(options);
   }
 }
