+++ src/apis/free-posts/free-posts.module.ts
@@ -1,7 +1,7 @@
 import { Module } from '@nestjs/common';
 import { CommonPostsModule } from '@src/apis/common-posts/common-posts.module';
+import { FreePostCommentRepository } from '@src/apis/free-post-comments/repositories/free-post-comment.repository';
 import { FreePostHistoryModule } from '@src/apis/free-posts/free-post-history/free-post-history.module';
-import { FreePostCommentRepository } from '@src/apis/free-posts/repositories/free-post-comment.repository';
 import { FreePostReplyCommentRepository } from '@src/apis/free-posts/repositories/free-post-reply-comment.repository';
 import { FreePostRepository } from '@src/apis/free-posts/repositories/free-post.repository';
 import { TypeOrmExModule } from '@src/core/type-orm/type-orm-ex.module';
@@ -22,5 +22,6 @@ import { FreePostsService } from './services/free-posts.service';
   ],
   controllers: [FreePostsController],
   providers: [FreePostsService, QueryHelper],
+  exports: [FreePostsService],
 })
 export class FreePostsModule {}
