+++ test/mock/mock.repository.ts
@@ -1,7 +1,6 @@
-import { FreePostCommentHistoryRepository } from '@src/apis/free-posts/free-post-history/repositories/free-post-comment-history.repository';
+import { FreePostCommentHistoryRepository } from '@src/apis/free-post-comments/free-post-comment-history/repositories/free-post-comment-history.repository';
+import { FreePostReplyCommentHistoryRepository } from '@src/apis/free-post-reply-comments/free-post-reply-comment-history/repositories/free-post-reply-comment-history.repository';
 import { FreePostHistoryRepository } from '@src/apis/free-posts/free-post-history/repositories/free-post-history.repository';
-import { FreePostReplyCommentHistoryRepository } from '@src/apis/free-posts/free-post-history/repositories/free-post-reply-comment-history.repository';
-import { FreePostCommentRepository } from '@src/apis/free-posts/repositories/free-post-comment.repository';
 import { FreePostReplyCommentRepository } from '@src/apis/free-posts/repositories/free-post-reply-comment.repository';
 import { FreePostRepository } from '@src/apis/free-posts/repositories/free-post.repository';
 import { MajorRepository } from '@src/apis/major/repositories/major.repository';
@@ -96,7 +95,7 @@ export const mockFreePostRepository: MockProvider<FreePostRepository> = {
   ...getDefaultRepositoryMethod(),
 };
 
-export const mockFreePostCommentRepository: MockProvider<FreePostCommentRepository> =
+export const mockFreePostCommentRepository: MockProvider<FreePostReplyCommentRepository> =
   {
     ...getDefaultRepositoryMethod(),
   };
